erp5_credential: use default_email_text related key to retrieve Person for Credential Recoveries
Previous behavior was to retrieve all emails, which can lead to
ununderstandable behaviors for the user. For exemple, if a user has 2
emails, one default, and one not, then in the credential recovery she
enters her second one, then a Credential Recovery is created, and the
recovery email is sent to her default email.